# Conflicts:
#	README.md

+++ src/main/java/games/pixscape/runtime/component/PixscapeIdentityComponent.java
@@ -3,7 +3,7 @@ package games.pixscape.runtime.component;
 import com.artemis.PooledComponent;
 
 public final class PixscapeIdentityComponent extends PooledComponent {
-    public long stableId = -1;
+    public int stableId = -1;
     public String name = "unnamed";
 
     @Override
